requirements.txt
mcp>=1.4.0
anthropic>=0.5.0
openai>=1.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
streamlit>=1.32.0